separate thread cards into it's own module
